middleware isloggedin, creation sauce + multer

+++ backend/controllers/sauce.js
@@ -1,8 +1,12 @@
 const Sauce = require('../models/Sauce');
 
 exports.createSauce = (req, res, next) => {
+    const sauceObject = JSON.parse(req.body.sauce); //createsauce fonctiuone sans isloggedqIN mais probleme id
+    //delete sauceObject.userId;
+
     const sauce = new Sauce({
-        ...req.body
+        ...sauceObject,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     sauce.save()
         .then(() => res.status(201).json({ message: 'Sauce enregistrée !' }))
@@ -10,20 +14,50 @@ exports.createSauce = (req, res, next) => {
 }
 
 exports.modifySauce = (req, res, next) => {
-    Sauce.updateOne({ _id: req.params.id }, {...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Sauce modifié !' }))
-        .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id }).then(
+        (sauce) => {
+            if (!sauce) {
+                return res.status(404).json({
+                    error: new Error('Sauce non trouvé !')
+                });
+            }
+            if (sauce.userId !== req.auth.userId) {
+                return res.status(401).json({
+                    error: new Error('Requete non autorisée !')
+                });
+            }
+            Sauce.updateOne({ _id: req.params.id }, {...req.body, _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Sauce modifié !' }))
+                .catch(error => res.status(400).json({ error }));
+        }
+    );
+
 }
 
 exports.deleteOneSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Sauce supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id }).then(
+        (sauce) => {
+            if (!sauce) {
+                return res.status(404).json({
+                    error: new Error('Sauce non trouvé !')
+                });
+            }
+            if (sauce.userId !== req.auth.userId) {
+                return res.status(401).json({
+                    error: new Error('Requete non autorisée !')
+                });
+            }
+            Sauce.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Sauce supprimé !' }))
+                .catch(error => res.status(400).json({ error }));
+        }
+    );
+
 }
 
 exports.showOneSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
-        .then(sauces => res.status(200).json(sauces))
+        .then(sauce => res.status(200).json(sauce))
         .catch(error => res.status(400).json({ error }));
 }
 exports.showAllSauces = (req, res, next) => {
